don't convert frequency automaticaly

+++ include/ad9910.h
@@ -258,7 +258,7 @@ void ad9910_enable_parallel(int enable);
  */
 INLINE void ad9910_set_parallel(uint16_t port);
 
-void ad9910_set_frequency(uint8_t profile, double freq);
+void ad9910_set_frequency(uint8_t profile, uint32_t freq);
 void ad9910_set_amplitude(uint8_t profile, uint16_t ampl);
 void ad9910_set_phase(uint8_t profile, uint16_t phase);
 
@@ -266,7 +266,7 @@ void ad9910_set_phase(uint8_t profile, uint16_t phase);
  * configures the specified profile to emit a sine with constant frequncy
  * and amplitude.
  */
-void ad9910_set_single_tone(uint8_t profile, double freq, uint16_t amplitude,
+void ad9910_set_single_tone(uint8_t profile, uint32_t freq, uint16_t amplitude,
                             uint16_t phase);
 
 /**
